Rename cs.ts to ntypes.cs.
