+++ src/redux/csState.ts
@@ -1,5 +1,5 @@
 import { PV_CHANGED, ActionType, SUBSCRIBE, WRITE_PV } from "./actions";
-import { NType } from "../cs";
+import { NType } from "../ntypes";
 
 const initialState: CsState = {
   valueCache: {}
